+++ sp-filter-builder/sp-filter-builder.js
@@ -5,8 +5,8 @@ import './sp-filter-builder.less!';
 import '../sp-list-table/sp-list-table';
 import '../sp-form/sp-form';
 import '../sp-filter/sp-filter';
-import '../sp-form/fields/text-field/text-field';
-import '../sp-form/fields/select-field/select-field';
+import '../sp-form/fields/sp-text-field/sp-text-field';
+import '../sp-form/fields/sp-select-field/sp-select-field';
 
 export default Component.extend({
     tag: 'sp-filter-builder',