+++ sp-form/fields/sp-select-field/sp-select-field.js
@@ -1,0 +1,9 @@
+import Component from 'can-component';
+import template from './sp-select-field.stache';
+import ViewModel from './ViewModel';
+
+export default Component.extend({
+    tag: 'sp-select-field',
+    view: template,
+    ViewModel: ViewModel
+});